python-can-isotp/requirements.txt
python-can
can-isotp
